feat(resource): 使用 MongoDB 存储文件
- 新增 MongoFileService 类实现文件上传和下载功能- 修改 ApplicationChatService、BaseDocumentExtractNode 等类，使用 MongoFileService 替代 FileService
- 更新 ChatFile 类，增加上传时间字段
- 重构 FileController，支持通过 HTTP 响应直接返回文件内容

+++ src/main/java/com/tarzan/maxkb4j/module/resource/service/FileService.java
@@ -32,7 +32,7 @@ public class FileService extends ServiceImpl<FileMapper, FileEntity>{
         vo.setFileId(fileEntity.getId());
         vo.setName(fileName);
         vo.setUrl("/api/file/"+fileEntity.getId());
-        vo.setSize(fileBytes.length);
+        vo.setSize((long) fileBytes.length);
         return vo;
     }
